+++ packages/expo-constants/android/src/main/java/expo/modules/constants/ConstantsService.kt
@@ -68,8 +68,6 @@ open class ConstantsService(private val context: Context) : InternalModule, Cons
   // Just use package name in vanilla React Native apps.
   override fun getAppScopeKey(): String? = context.packageName
 
-  override fun getAppOwnership() = "guest"
-
   override fun getDeviceName(): String = Build.MODEL
 
   override fun getIsDevice() = !EmulatorUtilities.isRunningOnEmulator()