+++ packages/expo-constants/src/Constants.types.ts
@@ -11,20 +11,11 @@ import type {
   // @ts-ignore -- optional interface, will gracefully degrade to `any` not installed
 } from 'expo-manifests';
 
-// @needsAudit
 export enum AppOwnership {
-  /**
-   * It is a [standalone app](/classic/building-standalone-apps#building-standalone-apps).
-   */
-  Standalone = 'standalone',
   /**
    * The experience is running inside of the Expo Go app.
    */
   Expo = 'expo',
-  /**
-   * It has been opened through a link from a standalone app.
-   */
-  Guest = 'guest',
 }
 
 // @docsMissing
@@ -128,9 +119,7 @@ export interface NativeConstants {
    */
   name: 'ExponentConstants';
   /**
-   * Returns `expo`, `standalone`, or `guest`. This property only applies to the managed workflow
-   * and classic builds; for apps built with EAS Build and in bare workflow, the result is
-   * always `null`.
+   * Returns `expo` when running in Expo Go, otherwise `null`.
    */
   appOwnership: AppOwnership | null;
   debugMode: boolean;